requirements.txt
asgiref==3.3.1
awscli==1.19.27
awsebcli==3.19.4
blessed==1.18.0
boto3==1.17.27
botocore==1.20.27
cached-property==1.5.2
cement==2.8.2
cffi==1.14.5
colorama==0.3.9
country-converter==0.7.2
cryptography==3.4.6
Django==3.1.5
django-rest-knox==4.1.0
django-storages==1.10
django-webpack-loader==0.6.0
djangorestframework==3.11.0
docker==3.7.3
docker-compose==1.25.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
jmespath==0.10.0
jsonschema==2.6.0
numpy==1.20.1
pandas==1.2.3
pathspec==0.5.9
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
s3transfer==0.3.4
semantic-version==2.8.5
six==1.11.0
sqlparse==0.4.1
tabulate==0.7.5
termcolor==1.1.0
texttable==0.9.1
urllib3==1.25.11
wcwidth==0.1.7
websocket-client==0.58.0
